fix(cli/secrets): Change order of fields in structs and move `get_app_id` to `utils`
