+++ lib/cli/src/commands/app/secrets/utils/mod.rs
@@ -1,8 +1,14 @@
 pub(crate) mod render;
 
 use colored::Colorize;
-use std::path::Path;
+use std::{
+    env::current_dir,
+    path::{Path, PathBuf},
+    str::FromStr,
+};
 use wasmer_api::{types::Secret as BackendSecret, WasmerClient};
+
+use crate::commands::app::util::{get_app_config_from_dir, prompt_app_ident, AppIdent};
 
 #[derive(serde::Serialize, serde::Deserialize)]
 pub(super) struct Secret {
@@ -80,5 +86,80 @@ pub(super) struct BackendSecretWrapper(pub BackendSecret);
 impl From<BackendSecret> for BackendSecretWrapper {
     fn from(value: BackendSecret) -> Self {
         Self(value)
+    }
+}
+
+/// A secrets-specific app to retrieve an app identifier.
+pub(super) async fn get_app_id(
+    client: &WasmerClient,
+    app: Option<&AppIdent>,
+    app_dir_path: Option<&PathBuf>,
+    quiet: bool,
+    non_interactive: bool,
+) -> anyhow::Result<String> {
+    if let Some(app_id) = app {
+        let app = app_id.resolve(client).await?;
+        return Ok(app.id.into_inner());
+    }
+
+    let path = if let Some(path) = app_dir_path {
+        path.clone()
+    } else {
+        current_dir()?
+    };
+
+    if let Ok(r) = get_app_config_from_dir(&path) {
+        let (app, _) = r;
+
+        let app_name = if let Some(owner) = &app.owner {
+            format!("{owner}/{}", app.name)
+        } else {
+            app.name.to_string()
+        };
+
+        let id = if let Some(id) = &app.app_id {
+            Some(id.clone())
+        } else if let Ok(app_ident) = AppIdent::from_str(&app_name) {
+            if let Ok(app) = app_ident.resolve(client).await {
+                Some(app.id.into_inner())
+            } else {
+                if !quiet {
+                    eprintln!("{}: the app found in {} does not exist.\n{}: maybe it was not deployed yet?", 
+                            "Warning".bold().yellow(), 
+                            format!("'{}'", path.display()).dimmed(), 
+                            "Hint".bold());
+                }
+                None
+            }
+        } else {
+            None
+        };
+
+        if let Some(id) = id {
+            if !quiet {
+                if let Some(owner) = &app.owner {
+                    eprintln!(
+                        "Managing secrets related to app {} ({owner}).",
+                        app.name.bold()
+                    );
+                } else {
+                    eprintln!("Managing secrets related to app {}.", app.name.bold());
+                }
+            }
+            return Ok(id);
+        }
+    } else if let Some(path) = app_dir_path {
+        anyhow::bail!(
+            "No app configuration file found in path {}.",
+            path.display()
+        )
+    }
+
+    if non_interactive {
+        anyhow::bail!("No app id given. Provide one using the `--app` flag.")
+    } else {
+        let id = prompt_app_ident("Enter the name of the app")?;
+        let app = id.resolve(client).await?;
+        Ok(app.id.into_inner())
     }
 }
